Move remote sampler to its own files
